requirements.txt
appdirs==1.4.4
astroid==2.12.13
autopep8==2.0.1
beautifulsoup4==4.11.1
boto3==1.26.109
botocore==1.29.109
bpemb==0.3.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
conllu==4.5.2
contourpy==1.0.6
cors==1.0.1
cryptography==40.0.1
cycler==0.11.0
Deprecated==1.2.13
dill==0.3.6
filelock==3.8.2
flair==0.11.3
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.38.0
frozendict==2.3.7
ftfy==6.1.1
future==0.18.2
gdown==4.4.0
gensim==4.2.0
gevent==22.10.2
greenlet==2.0.2
html5lib==1.1
huggingface-hub==0.11.1
hyperopt==0.2.7
idna==3.4
importlib-metadata==3.10.1
isort==5.11.4
itsdangerous==2.1.2
Janome==0.4.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
konoha==4.6.5
langdetect==1.0.9
lazy-object-proxy==1.8.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mccabe==0.7.0
more-itertools==9.0.0
mpld3==0.3
multiprocess==0.70.14
multitasking==0.0.11
networkx==2.8.8
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
overrides==3.1.0
packaging==22.0
pandas==2.0.0
Pillow==9.3.0
platformdirs==2.6.2
pptree==3.1
praw==7.6.1
prawcore==2.3.0
py4j==0.10.9.7
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.0.1
pylama==8.4.1
pylint==2.15.9
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
scikit-learn==1.2.0
scipy==1.9.3
segtok==1.5.11
sentencepiece==0.1.95
six==1.16.0
smart-open==6.2.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
sqlitedict==2.1.0
tabulate==0.9.0
threadpoolctl==3.1.0
tldextract==3.4.0
tokenizers==0.13.2
tomli==2.0.1
tomlkit==0.11.6
torch==1.13.0
tqdm==4.64.1
transformers==4.25.1
tweepy==4.12.1
typing_extensions==4.4.0
tzdata==2023.3
update-checker==0.18.0
urllib3==1.26.13
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
Wikipedia-API==0.5.4
wrapt==1.14.1
yfinance==0.2.15
zipp==3.11.0
zope.event==4.6
zope.interface==6.0
